fix: injected login bloc
