+++ lib/app/view/app.dart
@@ -1,23 +1,57 @@
+import 'package:authentication_repository/authentication_repository.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:fpb/authentication/application/bloc/authentication_bloc.dart';
 import 'package:fpb/core/shared/presentation/theming/themes/theme.dart';
 import 'package:fpb/l10n/l10n.dart';
 import 'package:fpb/router/go_routes.dart';
+import 'package:user_repository/user_repository.dart';
 
-class App extends StatelessWidget {
+class App extends StatefulWidget {
   const App({super.key});
 
   @override
+  State<App> createState() => _AppState();
+}
+
+class _AppState extends State<App> {
+  late final AuthenticationRepository _authenticationRepository;
+  late final UserRepository _userRepository;
+
+  @override
+  void initState() {
+    super.initState();
+    _authenticationRepository = AuthenticationRepository();
+    _userRepository = UserRepository();
+  }
+
+  @override
+  void dispose() {
+    _authenticationRepository.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return LayoutBuilder(
-      builder: (context, cts) {
-        return MaterialApp.router(
-          title: 'FinFlow',
-          theme: whiteTheme(context, cts),
-          localizationsDelegates: AppLocalizations.localizationsDelegates,
-          supportedLocales: AppLocalizations.supportedLocales,
-          routerConfig: goRouter,
-        );
-      },
+    return RepositoryProvider.value(
+      value: _authenticationRepository,
+      child: BlocProvider(
+        create: (context) => AuthenticationBloc(
+          authenticationRepository: _authenticationRepository,
+          userRepository: _userRepository,
+        ),
+        child: LayoutBuilder(
+          builder: (context, cts) {
+            return MaterialApp.router(
+              title: 'FinFlow',
+              theme: whiteTheme(context, cts),
+              localizationsDelegates: AppLocalizations.localizationsDelegates,
+              supportedLocales: AppLocalizations.supportedLocales,
+              routerConfig: goRouter,
+            );
+          },
+        ),
+      ),
     );
   }
 }
